server/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.10.6
sqlmodel==0.0.22
openai==1.66.3